+++ src/sdk/data/classes/quote.ts
@@ -1,0 +1,10 @@
+import { BigNumberish, BytesLike } from "ethers";
+
+export class Quote {
+    to?: string;
+    data?: BytesLike;
+    value?: BigNumberish;
+    gasLimit?: string;
+    gasPrice?: string;
+    chainId?: number;
+}