+++ src/sdk/data/data.module.ts
@@ -1,9 +1,10 @@
 import { BigNumber } from 'ethers';
 import { Route } from '@lifi/sdk';
 import { ObjectSubject } from '../common';
-import { AccountBalances, AdvanceRoutesLiFi, ExchangeOffer, NftList, PaginatedTokens, RateData, StepTransactions, TokenList, TokenListToken, Transaction, Transactions } from './classes';
+import { AccountBalances, AdvanceRoutesLiFi, Token, Quote, TransactionStatus, ExchangeOffer, NftList, PaginatedTokens, RateData, StepTransactions, TokenList, TokenListToken, Transaction, Transactions } from './classes';
 import { RestApiService } from '../api';
 import { API_ENDPOINTS, MethodTypes } from '../api/constants';
+import { BridgingProvider } from './constants';
 
 export class DataModule {
   readonly apiKey$ = new ObjectSubject<string>('');
@@ -247,6 +248,98 @@ export class DataModule {
       return result ? result.exchangeRates : null;
     } catch (error) {
       throw new Error(error.message || 'Failed to fetch exchange rates');
+    }
+  }
+
+  async getSupportedAssets(chainId?: number, provider?: BridgingProvider): Promise<Token[]> {
+    try {
+      const queryParams = {
+        'api-key': this.currentApi,
+        chainId,
+      };
+      let apiUrl: string;
+
+      switch (provider) {
+        case BridgingProvider.Connext:
+          apiUrl = API_ENDPOINTS.GET_CONNEXT_SUPPORTED_ASSETS;
+          break;
+        default:
+          apiUrl = API_ENDPOINTS.GET_CONNEXT_SUPPORTED_ASSETS;
+          break;
+      }
+
+      const result: { tokens: Token[] } = await this.apiService.makeRequest(apiUrl, MethodTypes.GET, queryParams);
+
+      return result ? result.tokens : [];
+    } catch (error) {
+      throw new Error(error.message || 'Failed to get supported assets');
+    }
+  }
+
+  async getQuotes(
+    fromAddress: string,
+    toAddress: string,
+    fromChainId: number,
+    toChainId: number,
+    fromToken: string,
+    fromAmount: BigNumber,
+    slippage: number,
+    provider?: BridgingProvider
+  ): Promise<Quote[]> {
+    try {
+      const queryParams = {
+        'api-key': this.currentApi,
+        fromAddress,
+        toAddress,
+        fromChainId,
+        toChainId,
+        fromToken,
+        fromAmount: fromAmount.toString(),
+        slippage
+      };
+      let apiUrl: string;
+
+      switch (provider) {
+        case BridgingProvider.Connext:
+          apiUrl = API_ENDPOINTS.GET_CONNEXT_QUOTE_TRANSACTIONS;
+          break;
+        default:
+          apiUrl = API_ENDPOINTS.GET_CONNEXT_QUOTE_TRANSACTIONS;
+          break;
+      }
+
+      const result: { transactions: Quote[] } = await this.apiService.makeRequest(apiUrl, MethodTypes.GET, queryParams);
+
+      return result ? result.transactions : [];
+    } catch (error) {
+      throw new Error(error.message || 'Failed to get quotes transactions');
+    }
+  }
+
+  async getTransactionStatus(fromChainId: number, toChainId: number, transactionHash: string, provider?: BridgingProvider): Promise<TransactionStatus> {
+    try {
+      const queryParams = {
+        'api-key': this.currentApi,
+        fromChainId,
+        toChainId,
+        transactionHash,
+      };
+      let apiUrl: string;
+
+      switch (provider) {
+        case BridgingProvider.Connext:
+          apiUrl = API_ENDPOINTS.GET_CONNEXT_TRANSACTION_STATUS;
+          break;
+        default:
+          apiUrl = API_ENDPOINTS.GET_CONNEXT_TRANSACTION_STATUS;
+          break;
+      }
+
+      const result: TransactionStatus = await this.apiService.makeRequest(apiUrl, MethodTypes.GET, queryParams);
+
+      return result ? result : null;
+    } catch (error) {
+      throw new Error(error.message || 'Failed to get transaction status');
     }
   }
 }